+++ src/errorHandler.js
@@ -1,45 +1,23 @@
-'use strict';
+"use strict";
 
-const printer = require('./shared/printer');
+const printer = require("./shared/printer");
 
-const TemplateNotFoundError = require('./errors/templateNotFoundError');
-const TemplateRenderingError = require('./errors/templateRenderingError');
-const TemplateInvalidError = require('./errors/templateInvalidError');
+const TemplateRenderingError = require("./errors/templateRenderingError");
 
-module.exports = (isVerbose, err) => {
-    if (err instanceof TemplateNotFoundError) {
-		printer.error('Template could not be found');
-
-		if (isVerbose) {
-			printer.error(err.innerError.message);
+function handleError(err) {
+	if (err instanceof TemplateRenderingError) {
+		printer.error("Template could not be rendered");
+		if (err.innerError) {
+			printer.errorDetails(err.innerError.message);
 		}
-
-		printer.info(`Try installing it with npm install -g ${err.templateName}`);
-
-	} else if(err instanceof TemplateRenderingError) {
-		printer.error('Template could not be rendered');
-
-		if (isVerbose) {
-			printer.error(err.innerError.message);
-		}
-
-		printer.info('The file has been created but left empty');
-
-	} else if (err instanceof TemplateInvalidError) {
-		printer.error('Template found but is invalid!');
-
-		if (isVerbose) {
-			printer.info(`The template ${err.template} does not expose a valid template function`);
-			printer.info(`If you are the author of the template then please check the documentation at https://github.com/dejanfajfar/tttemplate`)
-		}
+		printer.info("The file has been created but left empty");
 	} else {
-		printer.error('There was a catastrophic error!');
-
-		if (isVerbose) {
-			printer.error(err);
-			printer.error(err.stack);
-		}
+		printer.error("There was a catastrophic error!");
+		printer.errorDetails(err);
+		printer.errorDetails(err.stack);
 	}
 
 	process.exit();
-};
+}
+
+module.exports = handleError;