+++ src/parts/createFile.js
@@ -6,5 +6,5 @@ const printer = require("../shared/printer");
 module.exports = context => {
 	fsf.createFile(context.absolutePath);
 
-	printer.info(`Created empty file ${context.fullFileName}`);
+	printer.debug(`Created empty file ${context.fullFileName}`);
 };