+++ src/parts/printFileJobStartInfo.js
@@ -3,7 +3,7 @@
 const printer = require("../shared/printer");
 
 module.exports = context => {
-	printer.info(`File name ${context.fullFileName}`);
-	printer.info(`Target directory ${context.containingFolder}`);
-	printer.info(`Template to use ${context.template}`);
+	printer.debug(`File: ${context.fullFileName}`);
+	printer.debug(`Target: ${context.containingFolder}`);
+	printer.debug(`Template: ${context.template}`);
 };