+++ CHANGELOG.md
@@ -9,10 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - .ttouchrc file support added
 - debug output
+- errorDetails output
 
 ### Fixed
 
 - output consistency
+
+### Removed
+
+- Template invalid error
+- Template not found error
 
 ## [2.0.0]
 ### Added